requirements.txt
streamlit==1.32.0
beautifulsoup4==4.12.0
requests==2.31.0
pdfplumber==0.10.0
pandas==2.0.3
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
groq==0.3.0
unstructured==0.12.2
tiktoken==0.6.0
pdf2image==1.16.3
poppler-utils
pillow==10.1.0